seeding in progress

+++ db/migrate/20211122121250_create_transactions.rb
@@ -3,7 +3,7 @@ class CreateTransactions < ActiveRecord::Migration[6.1]
     create_table :transactions do |t|
       t.references :offer, null: false, foreign_key: true
       t.references :user, null: false, foreign_key: true
-      t.integer :item_price
+      t.float :item_price
 
       t.timestamps
     end
